requirements.txt
cbpro
pandas
numpy
prometheus_client
pyyaml
